+++ uamqp/__init__.py
@@ -35,7 +35,7 @@ except (SyntaxError, ImportError):
     pass  # Async not supported.
 
 
-__version__ = "1.5.3"
+__version__ = "1.6.0"
 
 
 _logger = logging.getLogger(__name__)