+++ HISTORY.rst
@@ -3,6 +3,16 @@
 Release History
 ===============
 
+1.6.0 (2022-08-18)
++++++++++++++++++++
+
+This version and all future versions will require Python 3.7+, Python 3.6 is no longer supported.
+
+- Added `data`, `value`, `sequence` properties to `uamqp.Message`, which return the body if the body type corresponds.
+- Added `message_annotations` property to `uamqp.Message`, which is an alias for the `annotations` instance variable.
+- Added `data` property to `uamqp.BatchMessage`, which returns the iterable body of the batch.
+- Added `ttl` property to `uamqp.MessageHeader`, which is an alias for the `time_to_live` instance variable.
+
 1.5.3 (2022-03-23)
 +++++++++++++++++++